prove that reduction do not change the result on working methods
this is achieved simply by copying the ¬reduction methods
in a withreduction method file. This way, the only
difference between naive_withreduction.lp and naive.lp
is that the first works on reduced graph.

Since both pass the same tests, we can assume that
the reduction method works correctly, and that any bug appearing on a
tierce withreduction method is due to the method, not the code
implementing the graph reduction.